+++ src/mailman/docs/NEWS.rst
@@ -26,6 +26,8 @@ Bugs
    a 409 error instead of a 500 error.  Found by Ankush Sharma.  (LP: #1425359)
  * ``mailman lists --domain`` was not properly handling its arguments.  Given
    by Manish Gill.  (LP: #1166911)
+ * When deleting a user object, make sure their preferences are also deleted.
+   Given by Abhishek.  (LP: #1418276)
 
 Configuration
 -------------